+++ r-backup.tf
@@ -1,7 +1,7 @@
 data "azurerm_subscription" "current" {}
 
 resource "null_resource" "backup" {
-  count = length(var.databases_names)
+  for_each = toset(var.databases_names)
 
   provisioner "local-exec" {
     command = <<EOC
@@ -11,14 +11,13 @@ resource "null_resource" "backup" {
       Connect-AzAccount -AccessToken $accessToken -AccountId $accountId -Force
       Get-AzSubscription -SubscriptionId ${data.azurerm_subscription.current.subscription_id} | Set-AzContext -Name "terraform-${data.azurerm_subscription.current.subscription_id}" -Force
 
-      Set-AzSqlDatabaseBackupShortTermRetentionPolicy -ResourceGroupName ${var.resource_group_name} -ServerName ${azurerm_sql_server.server.name} -DatabaseName ${element(var.databases_names, count.index)} -RetentionDays ${var.daily_backup_retention}
+      Set-AzSqlDatabaseBackupShortTermRetentionPolicy -ResourceGroupName ${var.resource_group_name} -ServerName ${azurerm_sql_server.server.name} -DatabaseName ${each.key} -RetentionDays ${var.daily_backup_retention}
 EOC
 
     interpreter = ["pwsh", "-c"]
   }
 
   triggers = {
-    database  = element(var.databases_names, count.index)
     retention = var.daily_backup_retention
   }
 
@@ -26,7 +25,7 @@ EOC
 }
 
 resource "null_resource" "ltr_backup" {
-  count = length(var.databases_names)
+  for_each = toset(var.databases_names)
 
   provisioner "local-exec" {
     command = <<EOC
@@ -37,7 +36,7 @@ resource "null_resource" "ltr_backup" {
       Get-AzSubscription -SubscriptionId ${data.azurerm_subscription.current.subscription_id} | Set-AzContext -Name "terraform-${data.azurerm_subscription.current.subscription_id}" -Force
 
       # Since this command is very long and can take around 1 hour, we do it asynchronously and wait a little to be sure it's started
-      Start-Job -ScriptBlock { Set-AzSqlDatabaseBackupLongTermRetentionPolicy -ResourceGroupName ${var.resource_group_name} -ServerName ${azurerm_sql_server.server.name} -DatabaseName ${element(var.databases_names, count.index)} -WeeklyRetention P${var.weekly_backup_retention}W -MonthlyRetention P${var.monthly_backup_retention}M -YearlyRetention P${var.yearly_backup_retention}Y -WeekOfYear ${var.yearly_backup_time} }
+      Start-Job -ScriptBlock { Set-AzSqlDatabaseBackupLongTermRetentionPolicy -ResourceGroupName ${var.resource_group_name} -ServerName ${azurerm_sql_server.server.name} -DatabaseName ${each.value} -WeeklyRetention P${var.weekly_backup_retention}W -MonthlyRetention P${var.monthly_backup_retention}M -YearlyRetention P${var.yearly_backup_retention}Y -WeekOfYear ${var.yearly_backup_time} }
       Start-Sleep 20
 EOC
 
@@ -45,7 +44,6 @@ EOC
   }
 
   triggers = {
-    database           = element(var.databases_names, count.index)
     weekly_retention   = var.weekly_backup_retention
     monthly_retention  = var.monthly_backup_retention
     yearly_retention   = var.yearly_backup_retention