+++ r-logs.tf
@@ -1,10 +1,8 @@
 module "logging" {
-  source = "github.com/claranet/terraform-azurerm-diagnostic-settings.git?ref=multiple-resources"
+  for_each = var.logs_destinations_ids != [] ? toset(var.databases_names) : toset([])
 
-  resources_count = var.enable_logging ? length(var.databases_names) : 0
-  resource_ids    = azurerm_sql_database.db[*].id
-  retention_days  = var.logs_storage_retention
+  source = "git::ssh://git@git.fr.clara.net/claranet/projects/cloud/azure/terraform/modules/diagnostic-settings.git?ref=AZ-160-revamp"
 
-  storage_account_id         = var.logs_storage_account_id
-  log_analytics_workspace_id = var.logs_log_analytics_workspace_id
+  resource_id           = azurerm_sql_database.db[each.key].id
+  logs_destinations_ids = var.logs_destinations_ids
 }