+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+Breaking
+  * AZ-273: Use for_each to iterate over databases
+  * AZ-160: Revamp diagnostic settings
+
 # v2.1.2/v3.0.1 - 2020-07-29
 
 Changed